pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sodsd"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Aleksandr" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "accelerate>=1.11.0",
    "black",
    "diffusers>=0.35.2",
    "flake8",
    "isort",
    "mkdocs",
    "numpy-stl>=3.2.0",
    "opencv-python>=4.12.0.88",
    "pip",
    "python-dotenv",
    "pyvista>=0.46.4",
    "stl>=0.0.3",
    "transformers>=4.57.1",
]
requires-python = "~=3.10.0"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["sodsd"]
force_sort_within_sections = true


